Windows:  Allow maximized windows to resize to fill the screen, even when a constrainer is applied
This mirrors the behaviour on Linux, where maximizing a window always
fills the screen, even if a constrainer is active.

If you don't want to allow your component to fill the screen, disable
the maximize button in the window's titlebar when creating the window
(i.e. ensure DocumentWindow::maximiseButton is not set in the required
buttons).