Improve SVG text parsing capabilities
Prior to this change all <tspan> elements without x, and y attributes would
just inherit the parent elements such attributes and be placed in the same
location. This didn't respect whether these attributes were consumed already
by the parent.

Having multiple x and y elements, or having a different number of x and y
elements was also not handled in line with the rules for SVG.