AU Client: Report correct AUChannelInfo for MIDI FX plugins
Previously, MIDI FX would report an input and output channel count of 0.
However, a non-empty output bus is required in order to retrieve the
processing sample rate.

With this change in place, MIDI FX plugins will now report an output
channel count of -1, which indicates that any number of output channels
is valid.