Direct2D: Reinstate multithreading support for D2D factory
Partially reverts a change made in
362a1cc07034655d2ff3dee786b3c8a72ea1145e.

The factory needs to be multithreading-enabled in order to support e.g.
drawing to the screen on the main thread and drawing to an Image on a
background thread, even when no resources other than the graphics device
are shared between the two threads.