PNGImageFormat: Avoid crashing when attempting to write an invalid image
The default error handler could cause crashes, so we now set up a custom
error handler both when reading and writing PNGs.

The HeapBlock and BitmapData automatic variables have moved, so that
their destructors will still run as expected in the failure case. Note
that it's UB to call longjmp to unwind the stack to the previous setjmp,
if said unwinding would normally cause non-trivial destructors to run.