MacOS: Fix WebBrowserComponent going blank in FL Studio
The issue could be triggered by opening the plugin in FL Studio, and
then using the TAB button to switch between FL Studio UI elements, until
the plugin became invisible and then it became visible again. This would
cause the WebBrowserComponent to navigate to about:blank permanently.

This was caused by the component becoming invisible and visible again in
rapid succession. This triggered a navigation to about:blank. To
understand the root cause of this, some undocumented behaviour of
WkWebView had to be uncovered. To understand this, see the following
test code, where the test1, test2 and test3 functions are called with
ample time in between one after the other.

void test1()
{
    goToURL ("A");
}

void test2()
{
    goToURL ("B");
    goToURL ("C");

    // B, C ignored completely, only D inserted into back-forward navigation queue
    goToURL ("D");
}

void test3()
{
    goToURL ("E");
    goToURL ("F");

    // E, F ignored completely, back navigation executed from D to A
    goBack();
}