VST Host: Use a "wrapper" window on Linux so that the client does not have to be responsible for window position
Previously, on Linux, client plugin editors were embedded directly into the peer
displaying the client's AudioProcessorEditor.

Although this approach is simple and lightweight, it means that plugin
editors are able to reposition themselves over controls in the parent
window just by calling XMoveResizeWindow or similar on their own widget.

It's more desirable that the client editor should be clipped if it
attempts to draw outside the area of the AudioProcessorEditor.